CON-4832: unassign articles from local categories unassignes current category
If you unassign a product from a local category in Import window,
the article was unassigned from all it's local categories.
This is not the action the user would expect.
Now only the currently selected Category and all child categories got removed.
If the user wants to unassign all categories he can do it over
the edit-function of the Article