CON-4922 Refactor storefront usage in stream export
Storefront namespace is designed to be used only in shopware frontend,
but we need to fetch articles/detal by stream id during stream export in backend.
Problem is that ShopContext is using Shop entity and it will break current session in DB.
With this PR Storefront namespace is not used anymore as a dependency.